LayoutTests/svg/animations/additive-from-to-width-animation-expected.txt

SVG 1.1 dynamic animation tests


This tests multiple additive='sum' animations running at the same time

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS rect.width.animVal.value is 20
PASS rect.width.baseVal.value is 10
PASS rect.width.animVal.value is 60
PASS rect.width.baseVal.value is 10
PASS rect.width.animVal.value is 100
PASS rect.width.baseVal.value is 10
PASS rect.width.animVal.value is 100
PASS rect.width.baseVal.value is 10
PASS successfullyParsed is true

TEST COMPLETE
